Upgrade Gradle, fix inconsistent behaviors and crashes on older API levels and non-english locales

+++ app/src/main/java/com/jdevs/timeo/model/ActivityItem.kt
@@ -11,7 +11,7 @@ import org.threeten.bp.OffsetDateTime
 data class ActivityItem(
     override val id: String,
     val name: String,
-    val totalTime: Long,
+    val totalTime: Int,
     val lastWeekTime: Int,
     val creationDate: OffsetDateTime
 ) : ViewItem, Parcelable
