+++ app/src/main/java/com/jdevs/timeo/model/ProjectItem.kt
@@ -12,7 +12,7 @@ data class ProjectItem(
     override val id: String,
     val name: String,
     val description: String,
-    val totalTime: Long,
+    val totalTime: Int,
     val lastWeekTime: Int,
     val creationDate: OffsetDateTime
 ) : ViewItem, Parcelable
